requirements.txt
PyMuPDF
langchain
langchain-core
langchain-community
chromadb
sentence-transformers
langchain-openai
python-dotenv